# docs/openapi.yaml
openapi: '3.0.3'
info:
  title: MindAtlas API
  description: AI-powered mapping and organization web application API
  version: 0.1.0
  contact:
    name: MindAtlas Team
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token obtained from POST /api/auth/login.
        Tokens expire after 15 minutes; use /api/auth/refresh to renew.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        API key created via POST /api/keys.
        Used for programmatic/webhook access.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid email or password
      required: [error]
...
    Subscription:
      type: object
      properties:
        plan:
          type: string
        status:
          type: string
          enum: [active, canceled, past_due]
        current_period_end:
          type: string
          format: date-time
...
paths:
...
  /api/csv/import:
    post:
      tags: [csv]
      summary: Import items from CSV
      description: Bulk-imports items from an uploaded CSV file
      operationId: csvImport
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file to import
      responses:
        '200':
          description: Import results
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        row:
                          type: integer
                        message:
                          type: string
        '400':
          description: Invalid CSV format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/csv/export/items:
    get:
      tags: [csv]
      summary: Export items as CSV
      operationId: csvExportItems
      security:
        - BearerAuth: []
      responses:
        '200':
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/csv/export/maps:
    get:
      tags: [csv]
      summary: Export maps as CSV
      operationId: csvExportMaps
      security:
        - BearerAuth: []
      responses:
        '200':
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/billing/subscription:
    get:
      tags: [billing]
      summary: Get current subscription
      operationId: getSubscription
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current subscription details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/billing/subscribe:
    post:
      tags: [billing]
      summary: Subscribe to a plan
      operationId: subscribe
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                plan_id:
                  type: string
                payment_method_id:
                  type: string
              required: [plan_id, payment_method_id]
      responses:
        '201':
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          description: Invalid plan or payment method
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/billing/upgrade:
    post:
      tags: [billing]
      summary: Upgrade subscription
      operationId: upgradePlan
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                plan_id:
                  type: string
              required: [plan_id]
      responses:
        '200':
          description: Plan upgraded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          description: Invalid upgrade path
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/billing/downgrade:
    post:
      tags: [billing]
      summary: Downgrade subscription
      operationId: downgradePlan
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                plan_id:
                  type: string
              required: [plan_id]
      responses:
        '200':
          description: Downgrade scheduled for end of billing period
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          description: Invalid downgrade path
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/billing/cancel:
    post:
      tags: [billing]
      summary: Cancel subscription
      operationId: cancelSubscription
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Subscription canceled at period end
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  cancel_at:
                    type: string
                    format: date-time
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/billing/history:
    get:
      tags: [billing]
      summary: Get billing history
      operationId: billingHistory
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of invoices/transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        amount:
                          type: number
                        status:
                          type: string
                        date:
                          type: string
                          format: date-time
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/billing/payment-method:
    put:
      tags: [billing]
      summary: Update payment method
      operationId: updatePaymentMethod
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                payment_method_id:
                  type: string
              required: [payment_method_id]
      responses:
        '200':
          description: Payment method updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Invalid payment method
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/billing/usage:
    get:
      tags: [billing]
      summary: Get usage metrics
      description: Returns current usage against plan limits
      operationId: billingUsage
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current usage
          content:
            application/json:
              schema:
                type: object
                properties:
                  storage_used_mb:
                    type: number
                  storage_limit_mb:
                    type: number
                  ai_queries_used:
                    type: integer
                  ai_queries_limit:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...